fix: search content and active item
